% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_r_package.R
\name{add_r_functions}
\alias{add_r_functions}
\title{\code{add_r_functions}}
\usage{
add_r_functions(function_names)
}
\arguments{
\item{function_names}{a string vector of function names that you want to expose to end users.}
}
\description{
\code{add_r_functions}
}
