% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{add_requirements}
\alias{add_requirements}
\title{\code{add_requirements} add tool requirement to the galaxy tool xml file.}
\usage{
add_requirements(packages, versions)
}
\arguments{
\item{packages}{a string vector of packages that can be found in the anaconda repository.}

\item{versions}{a string vector of corresponding package versions.}
}
\description{
\code{add_requirements} add tool requirement to the galaxy tool xml file.
}
