% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{add_requirements_command_line}
\alias{add_requirements_command_line}
\title{\code{add_requirements_command_line} add tool requirement to the galaxy tool xml file.}
\usage{
add_requirements_command_line(packages, versions)
}
\arguments{
\item{packages}{a string vector of packages that can be found in the anaconda repository.}

\item{versions}{a string vector of corresponding package versions.}
}
\description{
\code{add_requirements_command_line} add tool requirement to the galaxy tool xml file.
}
