% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{add_tools}
\alias{add_tools}
\title{\code{add_tools} add tools to 'elagator/elastic_tool_wrappers_macros.xml'.}
\usage{
add_tools(tools)
}
\arguments{
\item{tools}{a string vector of tool names.}
}
\description{
\code{add_tools} add tools to 'elagator/elastic_tool_wrappers_macros.xml'.
}
