% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{add_tools_command_line}
\alias{add_tools_command_line}
\title{\code{add_tools_command_line} add tools to 'elagator/elastic_tool_wrappers_macros.xml'.}
\usage{
add_tools_command_line(tools)
}
\arguments{
\item{tools}{a string vector of tool names.}
}
\description{
\code{add_tools_command_line} add tools to 'elagator/elastic_tool_wrappers_macros.xml'.
}
