% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_tool_repository.R
\name{create_shed_yml}
\alias{create_shed_yml}
\title{\code{create_shed_yml} creates a \code{.shed.yml} file within the elagator directory.}
\usage{
create_shed_yml(name, owner, description, long_description = "",
  type = "unrestricted", categories = "")
}
\arguments{
\item{name}{a string used as ToolShed repository name.}

\item{owner}{ToolShed user account.}

\item{description}{a brief introduction to the tool.}

\item{long_description}{detailed information about this tool.}

\item{type}{tool type. Default is 'unrestricted'}

\item{categories}{a vector of strings to specify which categories this tool belongs to.}
}
\description{
\code{create_shed_yml} creates a \code{.shed.yml} file within the elagator directory.
}
