% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygator.R
\docType{package}
\name{dygator-package}
\alias{dygator}
\alias{dygator-package}
\title{dygator: Assist in developing DYnamic GAlaxy TOols}
\description{
This package helps in developing DYnamic GAlaxy TOols.
}
\author{
\strong{Maintainer}: Ming Chen \email{ming.chen0919t@gmail.com}

}
\keyword{internal}
