% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\docType{package}
\name{elagator-package}
\alias{elagator}
\alias{elagator-package}
\title{elagator: Assist in developing ELAstic GAlaxy TOols in R}
\description{
This package helps in developing ELAstic GAlaxy TOols in R.
}
\author{
\strong{Maintainer}: Ming Chen \email{ming.chen0919t@gmail.com}

}
\keyword{internal}
