% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygator.R
\name{init_dygator}
\alias{init_dygator}
\title{\code{init_dygator} initialize a dynamic galay tool repository}
\usage{
init_dygator()
}
\description{
\code{init_dygator} initialize a dynamic galay tool repository
}
