% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{init_elagator}
\alias{init_elagator}
\title{\code{init_elagator} initialize a elastic galay tool repository}
\usage{
init_elagator()
}
\description{
\code{init_elagator} initialize a elastic galay tool repository
}
