% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{init_elagator_command_line}
\alias{init_elagator_command_line}
\title{\code{init_elagator_command_line} initialize a elastic galay tool repository}
\usage{
init_elagator_command_line()
}
\description{
\code{init_elagator_command_line} initialize a elastic galay tool repository
}
