% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_r_package.R
\name{init_elagator_r_package}
\alias{init_elagator_r_package}
\title{\code{init_elagator_r_package} initialize a elastic galay tool repository}
\usage{
init_elagator_r_package()
}
\description{
\code{init_elagator_r_package} initialize a elastic galay tool repository
}
