% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{it_tool}
\alias{it_tool}
\title{\code{it_tool} specify id, name, and version for your tool.}
\usage{
it_tool(id = "tool_id", name = "tool name", version = "1.0.0")
}
\arguments{
\item{id}{a string used for tool id. Default is 'tool_id'.}

\item{name}{a string used for tool name. Default is 'tool name'}

\item{version}{a string used for tool version. Defualt is '1.0.0'}
}
\description{
\code{it_tool} specify id, name, and version for your tool.
}
