% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_tool_repository.R
\name{publish_tool}
\alias{publish_tool}
\title{\code{publish_tool} publish/update tool repository to Galaxy Test ToolShed or ToolShed.}
\usage{
publish_tool(toolshed = "testtoolshed")
}
\arguments{
\item{toolshed}{a string to indicate which ToolShed to publish to. Valid values include 'testtoolshed' and
'toolshed'.}
}
\description{
\code{publish_tool} publish/update tool repository to Galaxy Test ToolShed or ToolShed.
}
