% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygator.R
\name{reinit_dygator}
\alias{reinit_dygator}
\title{\code{reinit_dygator} overwrite the existing dynamic galaxy tool repository with a new one.}
\usage{
reinit_dygator()
}
\description{
\code{reinit_dygator} overwrite the existing dynamic galaxy tool repository with a new one.
}
