% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{reinit_elagator}
\alias{reinit_elagator}
\title{\code{reinit_elagator} overwrite the existing elastic galaxy tool repository with a new one.}
\usage{
reinit_elagator()
}
\description{
\code{reinit_elagator} overwrite the existing elastic galaxy tool repository with a new one.
}
