% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{reinit_elagator_command_line}
\alias{reinit_elagator_command_line}
\title{\code{reinit_elagator_command_line} overwrite the existing elastic galaxy tool repository with a new one.}
\usage{
reinit_elagator_command_line()
}
\description{
\code{reinit_elagator_command_line} overwrite the existing elastic galaxy tool repository with a new one.
}
