% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_r_package.R
\name{reinit_elagator_r_package}
\alias{reinit_elagator_r_package}
\title{\code{reinit_elagator_r_package} overwrite the existing elastic galaxy tool repository with a new one.}
\usage{
reinit_elagator_r_package()
}
\description{
\code{reinit_elagator_r_package} overwrite the existing elastic galaxy tool repository with a new one.
}
