% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygator.R
\name{remove_requirements}
\alias{remove_requirements}
\title{\code{remove_requirements} remove existing requirements}
\usage{
remove_requirements(requirements)
}
\arguments{
\item{requirements}{a vector of package names.}
}
\description{
\code{remove_requirements} remove existing requirements
}
