% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{remove_requirements_command_line}
\alias{remove_requirements_command_line}
\title{\code{remove_requirements_command_line} remove existing requirements}
\usage{
remove_requirements_command_line(requirements)
}
\arguments{
\item{requirements}{a vector of package names.}
}
\description{
\code{remove_requirements_command_line} remove existing requirements
}
