% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_r_package.R
\name{remove_requirements_r_package}
\alias{remove_requirements_r_package}
\title{\code{remove_requirements_r_package} remove existing requirements}
\usage{
remove_requirements_r_package(requirements)
}
\arguments{
\item{requirements}{a vector of package names.}
}
\description{
\code{remove_requirements_r_package} remove existing requirements
}
