% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator_command_line.R
\name{remove_tools_command_line}
\alias{remove_tools_command_line}
\title{\code{remove_tools} remove tools from 'elagator/elastic_tool_wrappers_macros.xml'.}
\usage{
remove_tools_command_line(tools)
}
\arguments{
\item{tools}{a string vector of tool names.}
}
\description{
\code{remove_tools} remove tools from 'elagator/elastic_tool_wrappers_macros.xml'.
}
