% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygator.R
\name{add_requirements}
\alias{add_requirements}
\title{\code{add_requirements} add tool requirement to the galaxy tool xml file.}
\usage{
add_requirements(df)
}
\arguments{
\item{df}{a data frame or a matrix with two columns: the first column are package names from anaconda
repository; the second column are corresponding package versions.}
}
\description{
\code{add_requirements} add tool requirement to the galaxy tool xml file.
}
