% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elagator.R
\name{remove_tools}
\alias{remove_tools}
\title{\code{remove_tools} remove tools from 'elagator/elastic_tool_wrappers_macros.xml'.}
\usage{
remove_tools(tools)
}
\arguments{
\item{tools}{a string vector of tool names.}
}
\description{
\code{remove_tools} remove tools from 'elagator/elastic_tool_wrappers_macros.xml'.
}
